Manifest Changes
