+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Authentication/Login/MVP/LoginActivity.java
@@ -1,0 +1,6 @@
+package com.android.geeksNotNerds_Makearo.Authentication.Login.MVP;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class LoginActivity extends AppCompatActivity {
+}
