+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Authentication/Login/MVP/LoginContract.java
@@ -1,0 +1,4 @@
+package com.android.geeksNotNerds_Makearo.Authentication.Login.MVP;
+
+public class LoginContract {
+}
