+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Authentication/Login/MVP/LoginPresenter.java
@@ -1,0 +1,4 @@
+package com.android.geeksNotNerds_Makearo.Authentication.Login.MVP;
+
+public class LoginPresenter {
+}
