+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Authentication/Login/Model/LoginResponse.java
@@ -1,0 +1,4 @@
+package com.android.geeksNotNerds_Makearo.Authentication.Login.Model;
+
+public class LoginResponse {
+}
