+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Authentication/SignUp/MVP/SignUpContract.java
@@ -1,0 +1,4 @@
+package com.android.geeksNotNerds_Makearo.Authentication.SignUp.MVP;
+
+public class SignUpContract {
+}
