+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Authentication/SignUp/Model/SignUpResponse.java
@@ -1,0 +1,6 @@
+package com.android.geeksNotNerds_Makearo.Authentication.SignUp.Model;
+
+public class SignUpResponse
+{
+
+}
