+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Introduction/SplashScreen/SplashActivity.java
@@ -1,0 +1,8 @@
+package com.android.geeksNotNerds_Makearo.Introduction.SplashScreen;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class SplashActivity extends AppCompatActivity
+{
+
+}
