+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Introduction/SplashScreen/SplashContract.java
@@ -1,0 +1,4 @@
+package com.android.geeksNotNerds_Makearo.Introduction.SplashScreen;
+
+public class SplashContract {
+}
