+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Introduction/SplashScreen/SplashPresenter.java
@@ -1,0 +1,4 @@
+package com.android.geeksNotNerds_Makearo.Introduction.SplashScreen;
+
+public class SplashPresenter {
+}
