+++ app/src/main/java/com/android/geeksNotNerds_Makearo/MyFiles/MVP/MyFilesActivity.java
@@ -1,0 +1,4 @@
+package com.android.geeksNotNerds_Makearo.MyFiles.MVP;
+
+public class MyFilesActivity {
+}
