+++ app/src/main/java/com/android/geeksNotNerds_Makearo/Utilities/Notifications/MyFirebaseMessagingService.java
@@ -1,0 +1,4 @@
+package com.android.geeksNotNerds_Makearo.Utilities.Notifications;
+
+public class MyFirebaseMessagingService {
+}
