--- app/src/main/java/com/android/makearoapp/Authentication/Login/MVP/LoginActivity.java
@@ -1,6 +1,0 @@
-package com.android.makearoapp.Authentication.Login.MVP;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class LoginActivity extends AppCompatActivity {
-}
