--- app/src/main/java/com/android/makearoapp/Authentication/Login/MVP/LoginContract.java
@@ -1,4 +1,0 @@
-package com.android.makearoapp.Authentication.Login.MVP;
-
-public class LoginContract {
-}
