--- app/src/main/java/com/android/makearoapp/Authentication/Login/Model/LoginRequest.java
@@ -1,4 +1,0 @@
-package com.android.makearoapp.Authentication.Login.Model;
-
-public class LoginRequest {
-}
