--- app/src/main/java/com/android/makearoapp/Authentication/SignUp/MVP/SignUpContract.java
@@ -1,4 +1,0 @@
-package com.android.makearoapp.Authentication.SignUp.MVP;
-
-public class SignUpContract {
-}
