--- app/src/main/java/com/android/makearoapp/Authentication/SignUp/Model/SignUpResponse.java
@@ -1,6 +1,0 @@
-package com.android.makearoapp.Authentication.SignUp.Model;
-
-public class SignUpResponse
-{
-
-}
