--- app/src/main/java/com/android/makearoapp/Introduction/SplashScreen/SplashActivity.java
@@ -1,8 +1,0 @@
-package com.android.makearoapp.Introduction.SplashScreen;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class SplashActivity extends AppCompatActivity
-{
-
-}
