--- app/src/main/java/com/android/makearoapp/Introduction/SplashScreen/SplashContract.java
@@ -1,4 +1,0 @@
-package com.android.makearoapp.Introduction.SplashScreen;
-
-public class SplashContract {
-}
