--- app/src/main/java/com/android/makearoapp/MyFiles/MVP/MyFilesPresenter.java
@@ -1,4 +1,0 @@
-package com.android.makearoapp.MyFiles.MVP;
-
-public class MyFilesPresenter {
-}
