--- app/src/main/java/com/android/makearoapp/Utilities/Notifications/MyFirebaseMessagingService.java
@@ -1,4 +1,0 @@
-package com.android.makearoapp.Utilities.Notifications;
-
-public class MyFirebaseMessagingService {
-}
